requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
rest-framework-simplejwt==0.0.2
rpds-py==0.29.0
sqlparse==0.5.4
typing_extensions==4.15.0
uritemplate==4.2.0
